Move json functions to df_checker.py
